Datagram Client/Server
This Java example uses the Datagram class to send data from one client
to another. The data is a Java String object. The program serializes the
String object and sends it to a server. The server deserializes the
object and displays it in the window.

Run the program on two separate machines. Point them to each other using
IP addresses and the same port number.